Moved from local jars to maven dependencies for TeamCity server-api and common-api libraries. Updated the grade wrapper and java version. Bumped version in plugin description.
Plugin stopped working when I updated to TeamCity 2017.2 , I recompiled with Java 8 and with updated server jars, but turns out the config was just lost. Resetting the config, such as post url, make it work again. these changes may or may not be needed for the latest version, but they are done now anyway so releasing it.